% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_drought_map.R
\name{get_drought_map}
\alias{get_drought_map}
\title{Title}
\usage{
get_drought_map(date = "current", aoi, path = NULL, display_options = NULL)
}
\arguments{
\item{date}{date of map to retrieve. Default is "current" which will return the most recent map.}

\item{aoi}{the area of interest to get the map for}

\item{path}{file path to save map to}

\item{display_options}{display options to customize map returned}
}
\value{
returns a plot in the R viewer when path is NULL. If a file path
is specified, the map will be saved to that location.
}
\description{
Title
}
\examples{
get_drought_map(date = "current", aoi = "CA", display_options = "none")
}
