% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/monitor_drought.R
\name{monitor_drought}
\alias{monitor_drought}
\title{Retrieve data from the U.S. Drought Monitor API}
\usage{
monitor_drought(
  aoi,
  var,
  start_date = NULL,
  end_date = NULL,
  year = NULL,
  dx_lower = 0,
  dx_upper = 100
)
}
\arguments{
\item{aoi}{The are of interest to call drought data for. Can be one of: "us"
for the entire United States, "conus" for the continental United States.
Selecting a state can be done by either specifying the numeric FIPS code
corresponding to the state or by using the state's two letter
abbreviation (ex: California can be selected with either: 06 or "CA"). A
specific county can be selected by specifying a character string representing
the county FIPS code (ex: "06019" for Fresno County, California).}

\item{var}{a character string representing the variables to return. Can be
either "cs" for comprehensive statistics or "wid" for weeks in drought.}

\item{start_date}{"a character string representing the start date to retrieve data for"}

\item{end_date}{"a character string representing the end date to retrieve data for"}

\item{year}{"a character string or numeric value for the year to retrieve data for.
\code{year} can be specified instead of start_date and end_date to retrieve data
for an entire year or range of years"}

\item{dx_lower}{An optional parameter specifying the lower bound for the
drought statistics that records must be above to be returned.}

\item{dx_upper}{An optional parameter specifying the uppper bound for the
drought statistics that records must be above to be returned.}
}
\value{
returns a tibble
}
\description{
Retrieve data from the U.S. Drought Monitor API
}
\examples{
data <- monitor_drought(aoi = "us",var = "cs",
                        year = 2021,dx_lower = 0,
                        dx_upper = 100)
data <- monitor_drought(aoi = "01001",var = "cs",
                       start_date = "1/1/2021",
                       end_date = "12/31/2021")
}
